% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlplotter-package.R
\docType{package}
\name{nhlplotter}
\alias{nhlplotter}
\alias{nhlplotter-package}
\title{nhlplotter: Plotting tools for NHL data}
\description{
\code{nhlplotter} is an R package for plotting and visualizing National Hockey
League data. It is currently a package skeleton with plotting interfaces to
be added over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rentosaijo.github.io/nhlplotter/}
  \item \url{https://github.com/RentoSaijo/nhlplotter}
  \item Report bugs at \url{https://github.com/RentoSaijo/nhlplotter/issues}
}

}
\author{
\strong{Maintainer}: Rento Saijo \email{rentosaijo0527@gmail.com} (\href{https://orcid.org/0009-0008-4919-7349}{ORCID}) [copyright holder]

}
